requirements.txt
streamlit>=1.28.0
pandas
plotly
gspread
oauth2client
python-dotenv
vnstock
selenium
beautifulsoup4
webdriver-manager
scikit-learn
imbalanced-learn
xgboost
gspread_dataframe
altair<5
requests
numpy
openpyxl
